# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kürsü API",
    "description": "Anonim ders ve akademisyen değerlendirme platformunun API'si. Zarflar, akışlar, maskeleme ve hata gövdeleri gibi bu şemadan okunamayan kurallar için `docs/api-contract.md`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/professors": {
      "get": {
        "tags": [
          "professors"
        ],
        "summary": "List Professors",
        "operationId": "list_professors_professors_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Hoca adında arama",
              "title": "Search"
            },
            "description": "Hoca adında arama"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Sayfa başına kayıt",
              "default": 50,
              "title": "Limit"
            },
            "description": "Sayfa başına kayıt"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Atlanacak kayıt sayısı",
              "default": 0,
              "title": "Offset"
            },
            "description": "Atlanacak kayıt sayısı"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProfessorListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/professors/{professor_id}": {
      "get": {
        "tags": [
          "professors"
        ],
        "summary": "Get Professor Detail",
        "operationId": "get_professor_detail_professors__professor_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "professor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Professor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/faculties": {
      "get": {
        "tags": [
          "faculties"
        ],
        "summary": "List Faculties",
        "operationId": "list_faculties_faculties_get",
        "parameters": [
          {
            "name": "university_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Üniversite ID (zorunlu)",
              "title": "University Id"
            },
            "description": "Üniversite ID (zorunlu)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fakülte adında arama",
              "title": "Search"
            },
            "description": "Fakülte adında arama"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Sayfa başına kayıt",
              "default": 50,
              "title": "Limit"
            },
            "description": "Sayfa başına kayıt"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Atlanacak kayıt sayısı",
              "default": 0,
              "title": "Offset"
            },
            "description": "Atlanacak kayıt sayısı"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FacultyResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "faculties"
        ],
        "summary": "Create Faculty",
        "operationId": "create_faculty_faculties_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacultyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/faculties/{faculty_id}": {
      "patch": {
        "tags": [
          "faculties"
        ],
        "summary": "Update Faculty",
        "operationId": "update_faculty_faculties__faculty_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "faculty_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Faculty Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacultyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "faculties"
        ],
        "summary": "Delete Faculty",
        "operationId": "delete_faculty_faculties__faculty_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "faculty_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Faculty Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseProfessorSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_name": {
            "type": "string",
            "title": "Course Name"
          },
          "course_code": {
            "type": "string",
            "title": "Course Code"
          },
          "term": {
            "type": "string",
            "title": "Term"
          },
          "average_teaching_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Teaching Score"
          },
          "average_difficulty_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Difficulty Score"
          },
          "average_fairness_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Fairness Score"
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_name",
          "course_code",
          "term",
          "average_teaching_score",
          "average_difficulty_score",
          "average_fairness_score",
          "review_count"
        ],
        "title": "CourseProfessorSummary"
      },
...
      "FacultyCreate": {
        "properties": {
          "university_id": {
            "type": "integer",
            "title": "University Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "university_id",
          "name"
        ],
        "title": "FacultyCreate"
      },
      "FacultyResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "university_id": {
            "type": "integer",
            "title": "University Id"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "university_id"
        ],
        "title": "FacultyResponse"
      },
      "FacultyUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "university_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "University Id"
          }
        },
        "type": "object",
        "title": "FacultyUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_FacultyResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FacultyResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[FacultyResponse]"
      },
      "Page_ProfessorListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProfessorListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ProfessorListItem]"
      },
...
      "ProfessorDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseProfessorSummary"
            },
            "type": "array",
            "title": "Courses"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            },
            "type": "array",
            "title": "Reviews"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "courses",
          "reviews"
        ],
        "title": "ProfessorDetail"
      },
      "ProfessorListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "course_count": {
            "type": "integer",
            "title": "Course Count"
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count"
          },
          "average_teaching_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Teaching Score"
          },
          "average_difficulty_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Difficulty Score"
          },
          "average_fairness_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Fairness Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "course_count",
          "review_count",
          "average_teaching_score",
          "average_difficulty_score",
          "average_fairness_score"
        ],
        "title": "ProfessorListItem"
      },
...
      "ReviewResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "course_professor_id": {
            "type": "integer",
            "title": "Course Professor Id"
          },
          "teaching_score": {
            "type": "integer",
            "title": "Teaching Score"
          },
          "difficulty_score": {
            "type": "integer",
            "title": "Difficulty Score"
          },
          "fairness_score": {
            "type": "integer",
            "title": "Fairness Score"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_professor_id",
          "teaching_score",
          "difficulty_score",
          "fairness_score",
          "comment",
          "status",
          "created_at"
        ],
        "title": "ReviewResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}